+++ test/createProduct.test.js
@@ -1,4 +1,4 @@
-import { handler } from "../src/createProduct";
+import { handler } from "../src/productService/createProduct";
 
 const mockedProduct = { productId: 1 };
 
@@ -6,7 +6,7 @@ jest.mock("../src/utils/envUtils", () => ({
   checkRequiredEnvValues: jest.fn(),
 }));
 
-jest.mock("../src/productService/productRepository", function () {
+jest.mock("../src/productService/productModule/productRepository", function () {
   return () => ({
     createProduct: jest.fn(() => mockedProduct),
   });
