+++ test/productsById.test.js
@@ -1,5 +1,5 @@
-import { handler } from "../src/productsById";
-import ProductRepository from "../src/productService/productRepository";
+import { handler } from "../src/productService/productsById";
+import ProductRepository from "../src/productService/productModule/productRepository";
 
 const mockedProduct = { id: "1ac5310f-1025-4196-b49e-5188d1f0f63c" };
 
@@ -7,7 +7,7 @@ jest.mock("../src/utils/envUtils", () => ({
   checkRequiredEnvValues: jest.fn(),
 }));
 
-jest.mock("../src/productService/productRepository");
+jest.mock("../src/productService/productModule/productRepository");
 
 describe("productById", () => {
   beforeAll(() => {
